feat(Position): 공의 위치 값을 검증하는 로직 구현
