+++ docs/README.md
@@ -5,8 +5,8 @@
   - [x] **[예외 처리]** 공의 숫자가 1부터 9까지의 범위를 벗어나면 예외가 발생한다.
 
 ## Position
-- [ ] 공의 위치 값을 부여할 수 있다.
-  - [ ] **[예외 처리]** 공의 위치가 1부터 3까지의 범위를 벗어나면 예외가 발생한다.
+- [x] 공의 위치 값을 부여할 수 있다.
+  - [x] **[예외 처리]** 공의 위치가 1부터 3까지의 범위를 벗어나면 예외가 발생한다.
 
 ## Ball
 - [x] 공의 숫자와 위치 정보를 부여할 수 있다.
